refactored mainactivity to launch activity
