+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
             android:label="@string/title_activity_navigation_drawer"
             android:theme="@style/Theme.LanguageFlashCard.NoActionBar" />
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".ui.launch.LaunchActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
