+++ app/src/main/java/com/example/languageflashcard/ui/launch/LaunchViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.languageflashcard.ui
+package com.example.languageflashcard.ui.launch
 
 import androidx.lifecycle.ViewModel
 import com.google.firebase.auth.FirebaseAuth
@@ -6,7 +6,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class MainViewModel @Inject constructor(private val firebaseAuth: FirebaseAuth) : ViewModel() {
+class LaunchViewModel @Inject constructor(private val firebaseAuth: FirebaseAuth) : ViewModel() {
 
     fun getCurrentUser() = firebaseAuth.currentUser
 }
